+++ src/App.js
@@ -2,7 +2,7 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import AppRoutes from './routers/AppRoutes';
 import './App.css';
 
-function App() {
+const App = () => {
   return (
     <div className="App" style={{backgroundImage: 'url(https://i.ibb.co/0ZbZ6Ws/White-Walkers.jpg)',}}>
       <Router>
